Update build system
The existing GNU autoconf/automake based build system has stood the test
of time well. It almost worked 22 years after its creation and that is
no small feat.

I really like the build process standardized by autoconf. It's easy to
build all programs that rely on it in the same way. Far easier than
building software that rely on CMake IMHO. On the other hand the syntax
of autoconf (not automake so much) is arcane. It's not something I look
forward to revisit.

Of the more recent build system developments, meson[^1] is a personal
favorite. It remains to be seen what longevity meson will have. Will it
be more than a footnote in 20 years?

A build system with known longevity is GNU Make[^2]. The make syntax is
bizarre, probably worse than autoconf/automake, but I happen to be
somewhat familiar with it. For this reason the new build system to last
another 20 years is written in plain GNU Make.

A less appealing trait of GNU Make is that it has broken backwards
compatibility on several occasions. This may pose a problem down the
line. Time will tell...

[^1]: https://mesonbuild.com/
[^2]: https://www.gnu.org/software/make/