Fix out of bounds memory corruption in shields erosion code
Fix out of bounds memory corruption when shields are hit by a bomb or
missile. I had to refactor the code quite a bit in order to relearn it.

Younger me was still fond of M68k assembly when writing this code; loops
counting down, incrementing pointer stores (instead of indexing) and
lots of implicit data dependencies. E.g. passing a raw pointer to a
function that is expected to do the right thing using magic contants
sprinkled throughout.

It was interresing revisiting this code.
Without the stress of work it's like solving a puzzle.

The bug was caused by disregarding that the shield is expanded when the
convolution matrix to smooth its edges is applied. The shields starts
out with a 4 pixels black border and the code assumed such a margin was
in place. However after convolution, the black border is only 2 pixels.
The quality code I alluded to earlier then gladly accessed data outside
the shield buffer.