Create thin SDL wrapper library
Move all SDL calls into a thin wrapper library in preparation for the
SDL 2.x port. The game only requires a very minimal subset of SDL
functionality and keeping it in one place makes it easier to understand
what is needed for porting to other "media libraries".

Also, an assortment of general turd polishing. There are so many magic
numbers in this old creation.