Prevent pulse audio foolery
Mix in some inaudible noise to prevent pulse audio from stop playing the
game audio stream. It seems that pulse audio stops playing audio if the
stream is silent (all zeroes) for a short while. In our case, the
condition is triggered with a close to full invaders armada that moves
slowly, and thus generates tick sounds below the threshold frequency.

It took some hours to figure this one out :-/